Move the sg_io_hdr functionality to FuBlockDevice
Another vendor (Algoltek) wants to do almost exactly the same thing, so move to
a common location to avoid copy pasting complicated code.
